refactor: add barrel files and apply
