+++ src/components/PaymentModal/AddCardStep/index.tsx
@@ -1,11 +1,13 @@
-import { Card } from '@/components/common/Card'
-import { BottomSheet } from '@/components/common/BottomSheet'
-import { CardDateInput } from '@/components/common/CardDateInput'
-import { CardNumberInput } from '@/components/common/CardNumberInput'
-import { CardPasswordInput } from '@/components/common/CardPasswordInput'
-import { Input } from '@/components/common/Input'
-import { MaxLengthNumberInput } from '@/components/common/MaxLengthNumberInput'
-import { useDialog } from '@/hooks/useDialog'
+import {
+  Card,
+  BottomSheet,
+  CardDateInput,
+  CardNumberInput,
+  CardPasswordInput,
+  Input,
+  MaxLengthNumberInput
+} from '@/components/common'
+import { useDialog } from '@/hooks'
 
 export const AddCardStep = () => {
   const { isOpen, close, open } = useDialog()
