+++ src/components/PaymentModal/index.ts
@@ -1,0 +1,3 @@
+export * from './AddCardStep'
+export * from './CardListStep'
+export * from './CompletedAddCardStep'
