+++ src/constants/app.ts
@@ -15,3 +15,9 @@ export const CARD_PASSWORD = {
   MAX_LENGTH: 1,
   LAST_INPUT_INDEX: 3
 }
+
+export const PAYMENT_STEPS = [
+  'add card',
+  'completed add card',
+  'card list'
+] as const
